refactor: extract aceclaw-learning module (batch 1: leaf self-improvement classes)
aceclaw-daemon had grown to 29K LoC / 89 files, the largest module by far
(bigger than aceclaw-core + aceclaw-security + aceclaw-memory combined).
Most of the bloat is the self-learning subsystem (28 files, more incoming
per the open self-learning issues), which has zero conceptual coupling to
the daemon's actual responsibilities (UDS listener, JSON-RPC, agent loop,
lifecycle).

This batch carves out the 7 leaf-level self-improvement files with only
core + memory dependencies -- no in-daemon coupling -- so the migration
is purely mechanical (package change + import-add at consumers, no code
restructuring). Two related files (PatternDetector, LearningValidationRecorder)
depend on classes still in daemon (AgentSession, LearningValidationStore,
ValidationGateEngine, AutoReleaseController, SkillRefinementEngine) and
are deferred to later batches that move those transitively.

Moved (daemon -> learning, package dev.aceclaw.daemon -> dev.aceclaw.learning):

  ErrorDetector                  (264 LoC) + ErrorDetectorTest         (15 tests)
  FailureSignalDetector          (187 LoC) + FailureSignalDetectorTest (9 tests)
  LearningExplanation             (44 LoC)
  LearningExplanationRecorder    (256 LoC)
  LearningExplanationStore        (67 LoC) + LearningExplanationStoreTest (2 tests)
  LearningSignalReview            (52 LoC)
  LearningSignalReviewStore       (93 LoC) + LearningSignalReviewStoreTest (2 tests)

Net: daemon down 963 LoC. Diff is +68 / -11 (git rename detection sees
the move correctly -- the apparent 1500-line delta is just the same content
re-attributed to the new module path).

Consumer imports added in:
  aceclaw-daemon: AceClawDaemon, DynamicSkillGenerator, LearningValidationRecorder,
                  SelfImprovementEngine, SkillMemoryFeedback, StreamingAgentHandler
  aceclaw-daemon tests: CandidatePipelineIntegrationTest, DynamicSkillGeneratorTest,
                  SelfImprovementEngineTest, SelfLearningPipelineTest,
                  SkillMemoryFeedbackTest, StreamingAgentHandlerPostLearningTest

aceclaw-learning depends only on aceclaw-core (Turn, ContentBlock, Message)
and aceclaw-memory (AutoMemoryStore, Insight, candidate types). Crucially:
NO daemon dependency -- the module can be exercised in unit tests without
booting the daemon, and a future test author writing a learning unit test
doesn't have to drag the whole daemon classpath in.

Test plan:

- aceclaw-learning:test passes (28 tests in moved unit tests)
- All daemon consumer tests pass on this branch:
    SelfImprovementEngineTest (24), DynamicSkillGeneratorTest (13),
    SkillMemoryFeedbackTest (7), StreamingAgentHandlerPostLearningTest (1),
    CandidatePipelineIntegrationTest (10)
- SelfLearningPipelineTest has 2 pre-existing failures (verified identical
  on main with the same assertion failures -- not introduced here)
- Full compile across 14 modules passes

Follow-up batches will move:
  - Pattern detection + learning maintenance (~10 files) once AgentSession
    is either inlined or pulled to core
  - Skill draft/refinement subsystem (~10 files)
  - Validation + auto-release controllers (~5 files)

Eventual target: aceclaw-daemon below 20K LoC, with learning, validation
and skill subsystems each in their own module.
