+++ aceclaw-daemon/src/main/java/dev/aceclaw/daemon/AceClawDaemon.java
@@ -21,6 +21,13 @@ import dev.aceclaw.infra.event.DeferEvent;
 import dev.aceclaw.infra.event.EventBus;
 import dev.aceclaw.infra.event.SchedulerEvent;
 import dev.aceclaw.infra.health.*;
+import dev.aceclaw.learning.ErrorDetector;
+import dev.aceclaw.learning.FailureSignalDetector;
+import dev.aceclaw.learning.LearningExplanation;
+import dev.aceclaw.learning.LearningExplanationRecorder;
+import dev.aceclaw.learning.LearningExplanationStore;
+import dev.aceclaw.learning.LearningSignalReview;
+import dev.aceclaw.learning.LearningSignalReviewStore;
 import dev.aceclaw.llm.LlmClientFactory;
 import dev.aceclaw.memory.AutoMemoryStore;
 import dev.aceclaw.memory.CandidateStateMachine;
