+++ aceclaw-daemon/src/main/java/dev/aceclaw/daemon/LearningValidationRecorder.java
@@ -1,5 +1,7 @@
 package dev.aceclaw.daemon;
 
+import dev.aceclaw.learning.LearningSignalReview;
+
 import java.nio.file.Path;
 import java.util.List;
 import java.util.Objects;
