+++ aceclaw-daemon/src/main/java/dev/aceclaw/daemon/SkillMemoryFeedback.java
@@ -2,6 +2,8 @@ package dev.aceclaw.daemon;
 
 import dev.aceclaw.core.agent.SkillMetrics;
 import dev.aceclaw.core.agent.SkillOutcome;
+import dev.aceclaw.learning.LearningExplanation;
+import dev.aceclaw.learning.LearningExplanationRecorder;
 import dev.aceclaw.memory.AutoMemoryStore;
 import dev.aceclaw.memory.MemoryEntry;
