+++ aceclaw-daemon/src/test/java/dev/aceclaw/daemon/CandidatePipelineIntegrationTest.java
@@ -4,6 +4,8 @@ import dev.aceclaw.core.agent.ToolMetrics;
 import dev.aceclaw.core.agent.Turn;
 import dev.aceclaw.core.agent.ContextEstimator;
 import dev.aceclaw.core.llm.*;
+import dev.aceclaw.learning.ErrorDetector;
+import dev.aceclaw.learning.FailureSignalDetector;
 import dev.aceclaw.memory.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
