+++ aceclaw-daemon/src/test/java/dev/aceclaw/daemon/DynamicSkillGeneratorTest.java
@@ -7,6 +7,9 @@ import dev.aceclaw.core.llm.ContentBlock;
 import dev.aceclaw.core.llm.Message;
 import dev.aceclaw.core.llm.StopReason;
 import dev.aceclaw.core.llm.Usage;
+import dev.aceclaw.learning.LearningExplanation;
+import dev.aceclaw.learning.LearningExplanationRecorder;
+import dev.aceclaw.learning.LearningExplanationStore;
 import dev.aceclaw.memory.Insight;
 import dev.aceclaw.memory.PatternType;
 import org.junit.jupiter.api.BeforeEach;
