+++ aceclaw-daemon/src/test/java/dev/aceclaw/daemon/SelfImprovementEngineTest.java
@@ -3,6 +3,11 @@ package dev.aceclaw.daemon;
 import dev.aceclaw.core.agent.ToolMetrics;
 import dev.aceclaw.core.agent.Turn;
 import dev.aceclaw.core.llm.*;
+import dev.aceclaw.learning.ErrorDetector;
+import dev.aceclaw.learning.FailureSignalDetector;
+import dev.aceclaw.learning.LearningExplanation;
+import dev.aceclaw.learning.LearningExplanationRecorder;
+import dev.aceclaw.learning.LearningExplanationStore;
 import dev.aceclaw.memory.AutoMemoryStore;
 import dev.aceclaw.memory.FailureType;
 import dev.aceclaw.memory.Insight;
