+++ aceclaw-learning/src/main/java/dev/aceclaw/learning/LearningExplanation.java
@@ -1,4 +1,4 @@
-package dev.aceclaw.daemon;
+package dev.aceclaw.learning;
 
 import java.time.Instant;
 import java.util.List;
