+++ aceclaw-learning/src/test/java/dev/aceclaw/learning/LearningSignalReviewStoreTest.java
@@ -1,4 +1,4 @@
-package dev.aceclaw.daemon;
+package dev.aceclaw.learning;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
